Feat/header: enhance header layout and styles with improved navigatio… (#31)
* Feat/header: enhance header layout and styles with improved navigation and responsiveness

* Update AI button icon size to medium

Changed the icon size from large to medium for better visual alignment and consistency with the design. This ensures a more balanced appearance across the UI.

* Add HamBurgerButtonComponent to project

Introduce a reusable `HamBurgerButtonComponent` with a button containing a hamburger icon. This component extends `BaseElement`, utilizes Dota Core, and is styled for responsive behavior.

* Simplify and enhance header component's HTML structure.

Refactored the header component by cleaning up the HTML structure and improving readability. Updated class names and adjusted spacing for better clarity and styling consistency. Replaced the hamburger button with a new `ham-burger-button` component.

* Add HamBurgerButtonComponent to utils index export

This update includes the HamBurgerButtonComponent in the utils index file to make it accessible for imports. Ensures consistency and completeness in exporting all utility components.

* Add HamBurgerButtonComponent to main imports

The HamBurgerButtonComponent is now included in the utility components section of the main file. This ensures proper functionality and integration of the component where needed.

---------

Co-authored-by: ayu-sh-kr <akjaiswal2003@gmail.com>